fixed guildScheduledEventDelete crashing due to an event cancellation. Problem: non-regsitered events were only ignored if registered events in memory were <1 (i.e. not).
Cause: Cancelling a non registered event while there is a registered event already in memory.

Explanation: Since events in memory >0, bot searched for this cancelled event to pull it's details from memory but since it's not actually regsitered it crashed

Fix: added try{}catch() statement to try to remove events from memory. If event was not obtained from memory it is ignored and function returns.